pyproject.toml
[tool.poetry]
name = "em-unit-4-parser-async"
version = "0.1.0"
description = ""
authors = ["Alexandra Goldobina <goldobina.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
dotenv = "^0.9.9"
requests = "^2.32.3"
xlrd = "^2.0.1"
black = "^25.1.0"
asyncpg = "^0.30.0"
aiohttp = "^3.11.18"
aiofiles = "^24.1.0"
greenlet = "^3.2.0"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
redis = "^6.0.0"
fastapi-cache2 = "^0.2.2"
isort = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
